Full Upgrade. Many improvements. But yet not fully documented.

+++ Audio/VorbisFile.d
@@ -50,21 +50,20 @@ protected:
 		_sFile.bits  = 16;
 		_sFile.bytes = 2;
 		_sFile.dataSize = cast(uint) ov_pcm_total(&oggFile, -1) * _sFile.bytes * pInfo.channels;
-		_sFile.channels = cast(short) pInfo.channels;
+		_sFile.channels = cast(ushort) pInfo.channels;
 		debug writefln("Allocate %d memory for Vorbis.", _sFile.dataSize);
 		_sFile.buffer = new byte[_sFile.dataSize];
 		
-		int current = 0;
-		int endian  = 0; // 0 for Little-Endian, 1 for Big-Endian
+		uint current = 0;
+		int endian   = 0; // 0 for Little-Endian, 1 for Big-Endian
 		int bitStream;
 		
-		uint bytes;
+		long bytes;
 		
 		while (current < _sFile.dataSize) { // because it may take several requests to fill our buffer
 			bytes = ov_read(&oggFile,
-			                _sFile.buffer[current .. $].ptr, 
-			                _sFile.dataSize - current, 
-			                endian, 2, 1, &bitStream);
+			                _sFile.buffer[current .. $].ptr,
+			_sFile.dataSize - current, endian, 2, 1, &bitStream);
 			
 			current += bytes;
 		}
